dedup: share the simple method-signature renderer
MethodClassificationScanner.FormatDecodedSignature and
ExtensionMethodScanner.FormatMethodSignature carried the same ~20-line loop that
renders "ReturnType Name(type name, ...)" by matching parameter names via
SequenceNumber, differing only in the extension method's leading "this ". Extract it to
Metadata.SignatureRenderer.RenderDecodedSignature(..., extensionThis) and route both
callers through it. Output is unchanged; the richer ApiSurfaceExtractor renderer
(nullability, ref/out/params, defaults) is intentionally left as the separate
full-surface path.
